requirements.txt
pillow~=10.2.0
Jinja2~=3.1.2
SQLAlchemy~=1.4.49
loguru~=0.7.2
requests~=2.31.0
parsel~=1.8.1
